requirement.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
distro==1.9.0
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Markdown==3.8.2
pydantic==2.11.7
pydantic_core==2.33.2
PyQt5>=5.15.2
PyQt5_sip>=12.10.0
python-dotenv==1.1.1
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
openai>=1.40.0
requests>=2.31.0
